Extract control flow instructions to instructions_control.go module
