+++ cpu.go
@@ -525,91 +525,6 @@ func (c *CPU) IZY() uint8 {
 	return 0
 }
 
-// --- Instruction Operations (Stubs) ---
-// All operate functions remain the same method signatures:
-// func (c *CPU) OpName() uint8 { ... }
-
-func (c *CPU) BRK() uint8 {
-	// Note: PC is already incremented once by Clock() to point after the $00 opcode.
-	// The 6502 pushes PC+2 relative to the opcode address.
-	// The extra PC++ here achieves that. If removed, PC+1 would be pushed.
-	c.PC++ // Make PC point to PC+2 relative to opcode fetch address
-
-	// Push PC onto stack (now PC+2)
-	c.push16(c.PC)
-
-	// Push status register onto stack
-	// Note: B flag is SET, U flag is SET in the pushed copy
-	c.setFlag(B, true)
-	c.setFlag(U, true)
-	// --- Push P BEFORE setting I flag ---
-	originalP := c.P                 // Capture P before setting I
-	c.push(uint8(originalP | B | U)) // Push original flags + B + U
-	// Push P with B and U set
-	// c.push(uint8(c.P)) // Original incorrect line
-
-	// Set Interrupt Disable flag AFTER push
-	c.setFlag(I, true)
-
-	// Clear B flag in processor's actual register (it was only set for the push)
-	c.setFlag(B, false)
-
-	// Load interrupt vector ($FFFE/F)
-	lo := uint16(c.read(0xFFFE))
-	hi := uint16(c.read(0xFFFF))
-	c.PC = (hi << 8) | lo
-
-	// Set interrupt state (BRK uses IRQ vector)
-	c.inInterrupt = true
-	c.interruptVector = 0xFFFE
-
-	// BRK takes 7 cycles total (base cycles handled by lookup table)
-	// This function returns *extra* cycles, which should be 0 here.
-	return 0
-}
-
-func (c *CPU) JMP() uint8 {
-	c.PC = c.addrAbs
-	return 0
-}
-
-func (c *CPU) JSR() uint8 {
-	c.PC--
-	c.push16(c.PC)
-	c.PC = c.addrAbs
-	return 0
-}
-
-// NOP - No Operation
-func (c *CPU) NOP() uint8 {
-	// Page cross handling is now done via PageCrossPenalty field in lookup table
-	return 0
-}
-
-func (c *CPU) RTI() uint8 {
-	c.P = Flags(c.pop())
-	c.P &^= B
-	c.P |= U
-	c.PC = c.pop16()
-
-	// Clear interrupt state
-	c.inInterrupt = false
-	c.interruptVector = 0
-
-	return 0
-}
-
-func (c *CPU) RTS() uint8 {
-	c.PC = c.pop16()
-	c.PC++
-	return 0
-}
-
-func (c *CPU) XXX() uint8 {
-	log.Printf("ERROR: Illegal opcode $%02X encountered at $%04X", c.opcode, c.PC-1)
-	return 1 // Return 1 to prevent potential infinite loops if Clock checks >=0
-}
-
 // --- Lookup Table Initialization ---
 // *Use method expressions for assignment*
 func (c *CPU) buildLookupTable() {
